Chore: Refactoring file structure SeaWind
- Перестройка структуру под Clean w/o Domain & Docker Compose и будущее развитие проекта, а также под фичи VS
- Добавление Directory.Build.props - единые правила задания namespace & TargetFramework для всех проектов решения
- Добавление Directory.Packages.props - единый реестра версий всех проектов решения
- Refatoring для всех файлов правильных namespace
- Refactoring описания проектов с учетом наличия Directory.Packages.props
- Добавления заготовок для будующих проектов Application & Infrastructure
