+++ global.json
@@ -1,6 +1,5 @@
 {
   "sdk": {
-    "version": "8.0.414",
-    "rollForward": "latestFeature"
+    "version": "8.0.401"
   }
 }
