+++ src/backend/Api/Filters/CustomExceptionFilter.cs
@@ -1,5 +1,5 @@
-﻿using Backend.Exceptions;
-using Backend.Models;
+﻿using Api.Exceptions;
+using Api.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using System.Net;
