+++ src/backend/Api/Models/WeatherForecast.cs
@@ -1,4 +1,4 @@
-﻿namespace Backend.Models
+﻿namespace Api.Models
 {
     public record WeatherForecast(DateOnly Date, int TemperatureC, string? Summary)
     {
