+++ src/backend/Infrastructure/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace Infrastructure;
+
+public class Class1
+{
+
+}
