+++ src/frontend/web/eslint.config.js
